sql: use bigint for userid

+++ userbot/modules/sql_helper/antiflood_sql.py
@@ -4,7 +4,7 @@ except ImportError:
     raise AttributeError
 import threading
 
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String
 
 DEF_COUNT = 0
 DEF_LIMIT = 0
@@ -14,7 +14,7 @@ DEF_OBJ = (None, DEF_COUNT, DEF_LIMIT)
 class FloodControl(BASE):
     __tablename__ = "antiflood"
     chat_id = Column(String(14), primary_key=True)
-    user_id = Column(Integer)
+    user_id = Column(BigInteger)
     count = Column(Integer, default=DEF_COUNT)
     limit = Column(Integer, default=DEF_LIMIT)
