+++ userbot/modules/sql_helper/warns_sql.py
@@ -4,13 +4,13 @@ except ImportError:
     raise AttributeError
 import threading
 
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, distinct, func
+from sqlalchemy import BigInteger, Boolean, Column, Integer, String, UnicodeText, distinct, func
 
 
 class Warns(BASE):
     __tablename__ = "warns"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     num_warns = Column(Integer, default=0)
     reasons = Column(UnicodeText)
